refactor : 등록 서비스 추상화
