+++ Api/src/main/java/com/dingdong/api/s3/controller/S3Controller.java
@@ -3,7 +3,7 @@ package com.dingdong.api.s3.controller;
 
 import com.dingdong.api.s3.controller.request.DeleteImageRequest;
 import com.dingdong.api.s3.controller.response.UploadImageResponse;
-import com.dingdong.infrastructure.s3.S3Service;
+import com.dingdong.infrastructure.s3.IS3Service;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
@@ -17,7 +17,7 @@ import org.springframework.web.multipart.MultipartFile;
 @RequiredArgsConstructor
 public class S3Controller {
 
-    private final S3Service s3Service;
+    private final IS3Service s3Service;
 
     @Operation(summary = "이미지 업로드")
     @PostMapping(path = "/image", consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
