+++ Infrastructure/src/main/java/com/dingdong/infrastructure/s3/IS3Service.java
@@ -1,0 +1,10 @@
+package com.dingdong.infrastructure.s3;
+
+
+import org.springframework.web.multipart.MultipartFile;
+
+public interface IS3Service {
+    String uploadImage(MultipartFile multipartFile);
+
+    void remove(String fileUrl) throws Exception;
+}
